Replace XML configuration with JSON
This commit is pretty big. It's probably a good place to look for bugs.
I added a GSON dependency. This is now the only non-testing dependency.

The new (GSON-based) configuration file parser is **not yet
implemented**.

The long-range goal of this commit is more than just adding unnecessary
dependencies, messing with the configuration file format, and breaking
the code. The idea here is to cleanly communicate all the information
which is needed for the search thread. In fact, the impetus for these
changes was the desire to have the necessary information to conduct the
main search operation.